requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
boto3==1.12.0
botocore==1.15.0
census==0.8.13
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.3
django-allauth==0.41.0
django-appconf==1.0.3
django-db==0.0.7
django-embed-video==1.3.1
django-imagekit==4.0.2
django-restframework==0.0.1
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jellyfish==0.5.6
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
newrelic==5.6.0.135
oauthlib==3.1.0
optional-django==0.1.0
packaging==20.1
pilkit==2.0
Pillow==7.0.0
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-dotenv==0.10.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
us==1.0.0
virtualenv==20.0.7
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.2
yelpapi==2.4.0
